pyproject.toml
[project]
name = "spiderkit"
version = "0.1.9"
description = "一个面向爬虫与数据处理场景的 Python 工具包, 覆盖加密解密, 数据存储, 异步下载, 字体解析和哈希工具"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.11"
authors = [{name = "Dawn", email = "3439972272@qq.com"}]
license = {text = "MIT"}
keywords = [
    "spider",
    "crawler",
    "scraping",
    "encryption",
    "cryptography",
    "rsa",
    "aes",
    "des",
    "download",
    "async",
    "m3u8",
    "font-parser",
    "anti-crawler",
    "hash",
    "storage",
    "data-processing"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security :: Cryptography",
    "Topic :: Utilities",
]
dependencies = [
    "aiofiles>=23.0.0",
    "aiohttp>=3.8.0",
    "ddddocr>=1.4.0",
    "fonttools>=4.40.0",
    "inflection>=0.5.1",
    "loguru>=0.7.0",
    "m3u8>=3.5.0",
    "pillow>=10.0.0",
    "pycryptodome>=3.18.0",
    "tqdm>=4.65.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/spiderkit"]

[tool.uv]
publish-url = "https://upload.pypi.org/legacy/"

[project.urls]
Homepage = "https://github.com/dawn0608/spiderkit"
